Added warning if relevant PDF viewer bug detected
As of now, pdf.js has a serious bug ( https://bugzilla.mozilla.org/show_bug.cgi?id=1633908 ) that fails to render isolated points (moveto closepath) with linecap =1. (They should render as circles of diameter linewidth.) This prevents the proper display of the point plots in this slideshow. With this new version of gmovie, a bright red warning is displayed on the first page if (and, hopefully, only if) the bug is present. It is possible that the warning will be displayed fleetingly even if the bug is fixed, but it shouldn't persist. The warning includes a suggestion to use Adobe Acrobat Reader or other compatible viewer.