Major overhaul of ia.html's TeX->javascript
Mostly to handle mod function "properly". We first replace the modulus with its absolute value. For integer modulus, we make sure that the result follows standard math conventions for integer dividend, always returning a value in the range [0,modulus). For non-integer dividend, we convert the dividend to a rational, and if every prime factor of the denominator is smaller than and relatively prime to the modulus, we replace the dividend with the (integer) product of the numerator and the modular inverse of the denominator, and then use standard mod. In all other cases, we return the value in the range (-modulus/2,modulus/2] that differs from the dividend by an integer multiple of the modulus.
A particular problem case was a--b mod c which must be interpreted as a - (-b mod c); note that this is not the same as a+b mod c. Exponentiation is different; in that case binary and unary signs can be combined into a single binary sign, avoiding the need for additional parens to keep javascript happy--a unary sign is not allowed to immediately precede an exponential subexpression.
Another problem case was a mod b mod c, to be interpreted as (a mod b) mod c.