Modified plotting in ia.html
A heuristic attempt is made to plot discontinuous functions more faithfully by not drawing long lines between consecutive sample points. This is done by simple thresholding, so false positives (when the function is not actually discontinuous but a line segment is not drawn because the sample points are too far apart) and false negatives (when the function is discontinuous but a line segment is still drawn because the sample points are still fairly close together) can both occur.