Significant update to ia.html
Added log and atan, including '_' to specify base of log, and ',' to terminate base of log or first of two args to atan. Note that atan can take one or two args, with one arg equivalent to second arg = 1. Whether or not a base is specified, log is terminated with ')' (but '(' is explicit only if base is not specified.
Added keyboard equivalents for pi, tau, theta, log, atan.
Conversions between internal representations and TeX representation have been overhauled.
Help text has been expanded.